+++ Krowi_AchievementFilter/CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 ## [9.0.2.13.0] - 
 ### Added
 - Legion Zones (90 achievements)
-- Most parts of the Achievement window are now skinned when ElvUI's Blizzard Skins and Achievements are selected (Compatibility #7)
+- The Achievement window is now skinned when ElvUI's Blizzard Skins and Achievements are selected (Compatibility #7)
+- Right Click Menu, Filter Menu and Popup Dialog for copying links are skinned when ElvUI's Blizzard Skins and Misc Frames are selected (Compatibility #7)
+- Options are skinned when ElvUI's Ace3 Skins are selected (Compatibility #7)
+- Tutorials are skinned when ElvUI's Blizzard Skins and Tutorials are selected (Compatibility #7)
+
+### Changed
+- Updated all tutorial images to have a fade to transparent instead of fade to black to better work with ElvUI's Blizzard Skins and Tutorials (Compatibility #7)
 
 ## [9.0.2.12.0] - 2021-02-15
 ### Added